Search engine works

+++ app/src/main/java/com/example/artsell/adapters/ViewAllAdapter.java
@@ -41,7 +41,7 @@ public class ViewAllAdapter extends RecyclerView.Adapter<ViewAllAdapter.ViewHold
         holder.name.setText(list.get(position).getName());
         holder.description.setText(list.get(position).getDescription());
         holder.rating.setText(list.get(position).getRating());
-        holder.price.setText(Integer.toString(list.get(position).getPrice()));
+        holder.price.setText(Double.toString(list.get(position).getPrice()));
 
 
         holder.itemView.setOnClickListener(new View.OnClickListener() {
